test(LottoTest): 로또 번호 개수 틀릴 시 예외 발생 테스트 작성
