+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,0 +1,37 @@
+package lotto.domain;
+
+import lotto.Lotto;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import java.util.Arrays;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class LottoTest {
+    @ParameterizedTest
+    @DisplayName("로또 번호의 개수가 6개가 아니면 예외가 발생한다.")
+    @ValueSource(strings = {"1,2,3,4,5,6,7", "1,2,3,4,5"})
+    void Lotto_sizeIsNot6_ThrowException(String lotto) {
+        List<Integer> numbers = Arrays.stream(lotto.split(","))
+                                    .map(Integer::parseInt)
+                                    .toList();
+
+        assertThatThrownBy(() -> new Lotto(numbers))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageStartingWith("[ERROR]")
+                .hasMessageContaining("6개여야 합니다");
+    }
+
+    @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
+    @Test
+    void 로또_번호에_중복된_숫자가_있으면_예외가_발생한다() {
+        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    // TODO: 추가 기능 구현에 따른 테스트 코드 작성
+}
